pump_fun_bot/requirements.txt
solana==0.30.2
web3==6.11.3
python-dotenv==1.0.0
requests==2.31.0
websockets<12.0,>=9.0
gql==3.4.1
aiohttp==3.11.10
base58==2.1.1
